Fix copy-n-paste error which could cause many syscalls to fail on amd64.
git-svn-id: svn://svn.valgrind.org/valgrind/trunk@3868

+++ coregrind/m_syscalls/syscalls-main.c
@@ -327,7 +327,7 @@ void getSyscallStatusFromGuestState ( /*OUT*/SyscallStatus*     canonical,
 #  else
 #  if defined(VGP_amd64_linux)
    VexGuestAMD64State* gst = (VexGuestAMD64State*)gst_vanilla;
-   Long              i   = (Int)gst->guest_RAX;
+   Long                i   = (Long)gst->guest_RAX;
    canonical->what = i >= -4095 && i <= -1  ? SsFailure  : SsSuccess;
    canonical->val  = (UWord)(canonical->what==SsFailure ? -i : i);
 #  else
